requirements.txt
requests==2.25.1
beautifulsoup4==4.9.3
python-multipart==0.0.5
python-dotenv==0.15.0
flask-restx==0.2.0
gunicorn==20.0.4
flask_cors
requests_oauthlib
previsionio
flask